feat(GUI): Add a config selection button, as well as a configuration panel

+++ dopplerview/input_output/user_config.py
@@ -4,16 +4,16 @@ from pathlib import Path
 import shutil
 import sys
 from importlib.resources import files
-
+from importlib.metadata import version
 
 
 def get_user_config_dir():
     if sys.platform == "win32":
-        base = Path(os.getenv("LOCALAPPDATA"))
+        base = Path(os.getenv("APPDATA"))
     else:
         base = Path.home() / ".config"
 
-    path = base / "DopplerView"
+    path = base / "DopplerView" / version("dopplerview")
     path.mkdir(parents=True, exist_ok=True)
     return path
 
@@ -37,12 +37,22 @@ def get_latest_config(user_config, default_config):
         return default_config
     return user_config
 
-def ensure_config_file(filename):
+def ensure_latest_DV_config(config_path):
+    default_config = get_resource_path("default_dv_params.json")
+    user_config = ensure_config_file("DV_params.json", versioning=True)
+    latest_config = get_latest_config(user_config, default_config)
+
+    if latest_config != user_config:
+        shutil.copy(latest_config, user_config)
+
+    return latest_config
+
+def ensure_config_file(filename, versioning=False):
     user_dir = get_user_config_dir()
     user_file = user_dir / filename
 
     default_file = get_resource_path(filename)
-    if not user_file.exists() or get_version(user_file) < get_version(default_file):
+    if not user_file.exists() or (versioning and get_version(user_file) < get_version(default_file)):
         shutil.copy(default_file, user_file)
 
     return user_file
